Update of the client connect method
As the IP and port number of the server to connect the client to is
provided as a single string via the command line flag value, the
respective method for creating the client instance and connecting it to
the desired server should accept that single string as input parameter
only. The actual extraction of the IP and port number is part of that
methods now.