Addition of ComAnI default setting
In order to integrate the Commit Analysis Infrastructure (ComAnI) for
parsing commits received from a client in the server variant of
KernelHaven and for analyzing its introduced changes, that
infrastructure requires its own properties file. The new default setting
enables the definition of the path to that properties file as part of
the KernelHaven properties file passed as command line argument. Hence,
that path becomes part of the internal configuration instance of
KernelHaven and will be available for loading ComAnI.