Addition of IncrementalSetup
This class provides the ComAnI configuration (and maybe other
configurations related to incremental KernelHaven in future). That
configuration is used to instantiate the ComAnI infrastructure and its
plug-ins.
We could have reused the setup of ComAnI, but it has too restrictive
checks, creates directories, which are not used in this integration, and
prints ComAnI-specific information to the console, which are irritating.